Fix CI: split eligibility module; update admin-console e2e for new gate
Build failure
- Route components import ALLOWED_LEVELS / parseLevel / Level for JSX, but
  the module was named eligibility.server.ts. React Router's .server
  treeshake forbids client imports of *.server files, which broke both the
  client build (Build API) and the dev server used by Playwright (e2e-api).
- Split into:
    app/admin-console/lib/eligibility.ts          (client-safe: types, const, parser)
    app/admin-console/lib/eligibility.server.ts   (server-only: Prisma writers)
  Routes import client-safe items from the first and writers from the second.

E2E failure
- domain-management.spec "hiring lead cannot access admin console" was
  pinned to the pre-PR rule. Under the new tier, Core (Hiring Lead) is
  allowed in admin-console/members and admin-console/domains. Rewrote the
  describe block to assert the new contract: Core can reach members; a
  non-Core, non-Admin lab member (reviewer1) is still redirected.
