+++ dali-api/app/admin-console/lib/eligibility.server.ts
@@ -1,14 +1,17 @@
 import { prisma } from "~/lib/db";
+import type { Level } from "./eligibility";
 
-export type Level = "P1" | "P2" | "P3";
-
-// Shared writers for DomainEligibility used by:
+// Server-only writers for DomainEligibility used by:
 //   - admin-console/domains (per-domain Members cluster)
 //   - members/$id profile (Domains & Levels section)
 //
 // promotedBy is stamped on every upsert/update so the audit trail follows
 // the human who clicked, not the system (which is what staffing-finalize
 // and decision-release already do).
+//
+// Client-safe items (Level type, ALLOWED_LEVELS, parseLevel) are in
+// ./eligibility.ts — keep them imported from there in route components so
+// React Router's *.server treeshake doesn't trip on client code.
 
 export async function addOrUpdateEligibility(args: {
   userId: string;
@@ -32,11 +35,4 @@ export async function addOrUpdateEligibility(args: {
 
 export async function removeEligibility(args: { id: string }) {
   await prisma.domainEligibility.deleteMany({ where: { id: args.id } });
-}
-
-export const ALLOWED_LEVELS: Level[] = ["P1", "P2", "P3"];
-
-export function parseLevel(raw: unknown): Level | null {
-  if (typeof raw !== "string") return null;
-  return (ALLOWED_LEVELS as string[]).includes(raw) ? (raw as Level) : null;
 }
