+++ dali-api/e2e/domain-management.spec.ts
@@ -40,9 +40,21 @@ test.describe('admin domain management', () => {
   });
 });
 
-test.describe('non-admin domain management', () => {
-  test('hiring lead cannot access admin console', async ({ loginAs, page }) => {
+test.describe('admin console access tiers', () => {
+  // Core (Hiring Lead title) gained access to admin-console/members and
+  // admin-console/domains so they can manage Core titles, Domain Leads, and
+  // member eligibilities without needing Admin. Admin-only actions
+  // (set-admin, create-domain, delete-domain) still 403 inside the action
+  // handler.
+  test('hiring lead (Core) can access admin console members', async ({ loginAs, page }) => {
     await loginAs({ daliEmail: 'jordan.taylor@dali.dartmouth.edu' });
+    await page.goto('/admin-console/members');
+    await expect(page).toHaveURL(/\/admin-console\/members/);
+  });
+
+  // A lab member with neither Core nor Admin is still redirected away.
+  test('non-Core, non-Admin lab member cannot access admin console', async ({ loginAs, page }) => {
+    await loginAs({ daliEmail: 'reviewer1@dali.dartmouth.edu' });
     await page.goto('/admin-console/members');
     await expect(page).not.toHaveURL(/admin-console/);
   });
